+++ src/views/ImgResponsive.vue
@@ -1,5 +1,11 @@
 <template>
   <section>
+    <div data-grid="center">
+      <div data-col="10@md 8@lg">
+        <h2>responsiveness</h2>
+      </div>
+    </div>
+
     <div class="demo-preview has-padding-5 has-margin-bottom-5 is-relative">
       <div data-grid="center">
         <div data-col="4">
@@ -11,39 +17,27 @@
       <div
         class="demo-preview-settings is-absolute is-fab-s is-theme is-elevated-16 has-padding-middle-2 has-padding-center-3 is-round"
       >
-        <div data-grid>
-          <div data-col="6">
-            <div class="xui-radio">
-              <input
-                type="radio"
-                id="imgResponsive"
-                name="responsiveness"
-                value="img-responsive"
-                v-model="responsiveness"
-              >
-              <label for="imgResponsive" class="xui-radio__label">responsive</label>
-            </div>
+        <div class="nav">
+          <div class="nav__item">
+            <a
+              @click="responsiveness = 'is-responsive'"
+              class="btn btn_clear is-uppercase"
+              :class="{'is-ink': responsiveness !== 'is-responsive'}"
+            >is-responsive</a>
           </div>
-          <div data-col="6">
-            <div class="xui-radio">
-              <input
-                type="radio"
-                id="imgFull"
-                name="responsiveness"
-                value="img-full"
-                v-model="responsiveness"
-              >
-              <label for="imgFull" class="xui-radio__label">full</label>
-            </div>
+          <div class="nav__item">
+            <a
+              @click="responsiveness = 'is-fullwidth'"
+              class="btn btn_clear is-uppercase"
+              :class="{'is-ink': responsiveness !== 'is-fullwidth'}"
+            >is-fullwidth</a>
           </div>
         </div>
       </div>
     </div>
     <div data-grid="center">
-      <div data-col="8@xl">
-        <div class="demo-markup-preview has-margin-middle-5">
-          <pre v-highlightjs="markup"><code class="html"></code></pre>
-        </div>
+      <div data-col="10@md 8@lg">
+        <pre v-highlightjs="markup"><code class="html"></code></pre>
       </div>
     </div>
   </section>
@@ -59,8 +53,8 @@ export default {
   },
   data: () => {
     return {
-      responsiveness: 'img-responsive',
-      markup: `<img class="img-responsive" />`,
+      responsiveness: 'is-responsive',
+      markup: `<img class="is-responsive" />`,
     };
   },
   watch: {